+++ extensions/index.ts
@@ -2,3 +2,8 @@ export {
   initAligningGuidelines,
   type AligningLineConfig,
 } from './aligning_guidelines';
+
+export {
+  originUpdaterWrapper,
+  installOriginWrapperUpdater,
+} from './data_updaters/origins';